Fifth commit:
Program class - added main method call to switch statement for when a user decides to clear recipe. This is so they can enter a new recipe. Changed foreground colour of goodbye message to blue so it matches the welcome message. Included break statement inside if statement to exit the while loop if user decides to clear and enter a new recipe so that the option menu is not repeated when the user exits the application.