Sixteenth commit:
Program class - Got rid of recipe and display names as now the recipe name is stored in the recipes list and the recipes are numbered when they are displayed. So user can enter the number rather than the name... reduce user error. Added some comments on the delegate and event. Decided to create a new recipe object with a recipeID and recipeName. Ingredient objects and steps will be added to every newely created recipe object. Moved the total calorie check to right after the adding of a new recipe object. Moved the CalculateTotalCalories method to the Recipe class. Added totalCalories as a parameter to the RecipeExceededCaloriesHandler method.

Recipe class - Got rid of Program class object instantiation. Added RecipeID and RecipeName variables. Added said variables as parameters to Recipe constructor. Added recipe object of Recipe class as parameter to AddIngredient, AddStep, DisplayRecipe, and CalculateTotalCalories methods. Added statement to display total calories in the DisplayRecipe method using CalculateTotalCalories method. Got rid of GetIngredients method as it was not being called. Changed the sorting of the recipes list in the DisplayRecipeList method. CHanged the statement so that the recipe names are numbered.  Finished the DisplaySpecificRecipe method. Includes asking for the recipe number from the user. Hence the number of the recipe names in the DisplayRecipeList method. Next is checking the validity of the entered recipe number. Then finding the recipe corresponding with said number if it's valid. Then displaying recipe details. If recipe number is not valid, appropriate output statements will be executed.