Seventeenth commit:
Program class - assigned null value to clearConfirm and allowed for it to be nullable. Initialised recipe in Program constructor with a default value. Passed recipe object into ActionsMenu as an argument. Added while loop and if statement to continuosuly ask user for a valid input if the input given is null for recipe name, ingredient name, ingredient quantity, ingredient food group, step description, and clear confirm. Added Recipe recipe as a parameter to ActionsMenu method.

Recipe class - made recipes list static and put the instantiation at the top outside of the methods. Ammended DisplayRecipe method by putting recipe. before RecipeName, ingredients, and steps to ensure i was accessing the details of the correct recipe object. Added a statement to display total calories in teh DisplaySpecificRecipe method.