Eighteenth commit:
Program class - declared recipeIDCounter as static. Changed GetRecipeDetails method return type to Recipe so that it can return a Recipe object. As a result i added a return statement to the end of the method returning 'recipe' which is the new recipe added. Passed recipe as an argument to ScaleRecipe method.

Recipe class - added Recipe as a parameter to ScaleRecipe method. Also changed the loop from a standard for loop to a foreach loop in ScaleRecipe method.